+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - [Optimize: Register the service with the ProviderAPI#registerInstance method.](https://github.com/Tencent/spring-cloud-tencent/pull/686)
 - [docs:update PR template.](https://github.com/Tencent/spring-cloud-tencent/pull/690)
 - [enhancement: revert default value when the field is deleted in @ConfigurationProperties bean](https://github.com/Tencent/spring-cloud-tencent/issues/681)
+- [Feature:support multi register and discovery both to nacos and polaris.](https://github.com/Tencent/spring-cloud-tencent/pull/693)
 - [Code optimize & add junit tests.](https://github.com/Tencent/spring-cloud-tencent/pull/701)
 - [Optimize: remote deprecated method](https://github.com/Tencent/spring-cloud-tencent/pull/697)
 - [Test:support environment variable metadata test](https://github.com/Tencent/spring-cloud-tencent/pull/698)
+- [Test: remote deprecated method in test](https://github.com/Tencent/spring-cloud-tencent/pull/706)